Remove focus grabber from the tracklist.
When the user mouses over the track list, it will move it to the top of
the page. This is annoying if you're trying to click one of the controls
but slightly miss or inadvertently drag your mouse over the track list.
This means the cursor won't change as desired when over the tracklist and
pressing ENTER or CTRL, but that's fine.