Ensure the playlist page works after re-auth.
If the auth token expired and the user tries to load a specific playlist
page, then the auth dialog was preventing the bulk "Delete" and "Add To
Library" buttons from being hooked up.

Also, the playlist would be stick in multi-select mode, as if the user
had the Ctrl button held down.