Lazy loading of playlist duration.
When retrieving all your playlists from YouTube Music, duration info is
not returned. The only way to get it is retrieve all the tracks from the
playlist then sum their lengths. However, this call is very slow, and
must be made for each playlist separately, which leads to the user
staring at a blank screen for a while. And due to Heroku's 30 second
request timeout, it's not rare for the list to never load.

To solve this, we bump the playlistmanager version so it uses a light
listing when being asked for info on all your YTM playlists. Then, once
the whole list is displayed, we request detailed info for each playlist,
and add the duration to the list entry.

Note these follow up requests are forced to execute in serial. This is
to reduce the chance of getting a 429 from YTM due to rate limiting.